queue tasks using strings to avoid circular imports when attmepting to queue w/ burnettk
