+++ spiffworkflow-backend/src/spiffworkflow_backend/routes/process_instances_controller.py
@@ -13,10 +13,10 @@ from flask.wrappers import Response
 from sqlalchemy import or_
 from sqlalchemy.orm import aliased
 
-from spiffworkflow_backend.background_processing.celery_tasks.process_instance_task import (
+from spiffworkflow_backend.background_processing.celery_tasks.process_instance_task_producer import (
     queue_enabled_for_process_model,
 )
-from spiffworkflow_backend.background_processing.celery_tasks.process_instance_task import (
+from spiffworkflow_backend.background_processing.celery_tasks.process_instance_task_producer import (
     queue_process_instance_if_appropriate,
 )
 from spiffworkflow_backend.data_migrations.process_instance_migrator import ProcessInstanceMigrator
